Removed the isSingleton hack from the view
I had an issue with recursive object construction in the View method because I was putting the View component onto the JPanel by creating a new instance. I fixed this problem by passing `this` to the JPanel component, using the very same instance of the View rather than making a new one.