Add KeyAdapter and isMoving to model
Generally, the model really shouldn't be extending keyAdapter, it shouldn't "control" itself (that's what a controller is for), but the way Greg wrote the code means that we'll do it this way. Ideally the model would only contain the "state" of the application and not update itself. The KeyAdapter should be extracted into its own class and then utilize "setters" on the model.

This works for now so I'm going to leave it at that.