Removed redundant code, improved event-based proc.
There were spots where we set the same values multiple times throughout the lifecycle of the View, this resulted in strange flickering. I removed that. I also modified the update method to clear the panel before redrawing. I used button.repaint() to ensure the button wasn't obscured by the orc image drawing. I also set the action to idle if the orc is not moving.