Moved action changes into model
This way the model keeps the state of the actions, and the view just reflects that action by showing the correct animation files.

I also implemented the Jump button, but it is kinda buggy. There might be a better way to do it.
You can see that I've added a tick_counter that gets passed to the model update function. This way we can count how many ticks go by and ensure that the Jump animation is complete before changing the action to something else.

In the View, I changed the update method to check if the next action is different from the last action, and if it is, it should reset the picNum to 0 so that the animation starts from the beginning. This is turning out to be kinda buggy, and a better solution should be devised.

I also updated the OrcImage enum to provide some helper functions that let us call OrcImage.forward(dir), for instance. This means we can update the orc's direction once every loop and don't need to write obnoxious switch statements every time we want to do a different action.